refactor(cli): import RPC enums from notebooklm.types (T4.2)
Swap three CLI imports from ``notebooklm.rpc`` to the public
``notebooklm.types`` surface, which already re-exports these enums.
Keeps CLI inside the documented public boundary so the upcoming
``test_cli_boundary.py`` block-list (T4.1) can forbid ``..rpc``
imports from ``cli/``.

Also adds a parametrized identity check in
``tests/unit/test_public_shims.py`` so the re-exports are guaranteed
to be the same objects as the canonical enums in ``notebooklm.rpc.types``
(not just equal), keeping isinstance and equality usage safe across
import paths.
